+++ CHANGELOG.md
@@ -1,16 +1,17 @@
 # Change Log
 
-## v0.xx.x
+
+## v0.x.x Unreleased
 
 ### New features
 * Add `weight_predictions` function to allow generation of weighted predictions from two or more InfereceData with `posterior_predictive` groups and a set of weights ([2147](https://github.com/arviz-devs/arviz/pull/2147))
 
 ### Maintenance and fixes
+- Fix dimension ordering for `plot_trace` with divergences ([2151](https://github.com/arviz-devs/arviz/pull/2151))
 
 ### Deprecation
 
 ### Documentation
-
 
 ## v0.13.0 (2022 Oct 22)